requirements.txt
huggingface-hub
llama-cpp-python
pandas
scikit-learn